requirements.txt
# Core Data Analysis
pandas==2.1.4
numpy==1.26.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Data Generation
faker==20.1.0
python-dateutil==2.8.2

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
statsmodels==0.14.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Dashboard
streamlit==1.29.0

# Jupyter
jupyter==1.0.0
ipykernel==6.27.1
notebook==7.0.6

# Utilities
python-dotenv==1.0.0
openpyxl==3.1.2

# Reporting
reportlab==4.0.7
fpdf2==2.7.7
jinja2==3.1.2

# Testing & Validation
pytest==7.4.3
great-expectations==0.18.8



